Create README.md
Pneumonia detection will involve the use of deep learning techniques. Specifically, a convolutional neural network (CNN) will be trained on a large dataset of chest X-ray images to learn patterns and features that are indicative of pneumonia. The CNN will then be used to predict whether a given chest X-ray image contains pneumonia or not. To ensure the accuracy of the model, the dataset will be split into training, validation, and testing sets.